+++ docs/source/changelog.rst
@@ -12,4 +12,7 @@ Changelog
 ..
     For Help: https://releases.readthedocs.io/en/latest/index.html
 
+
+* :feature:`x` Landmark paper: The Distance Approach (Gatev et al. 2006).
+* :support:`x` Distance approach documentation.
 * :support:`10` Added a number of new tools to improve our deployment and align us with best practices. They include: CircleCI, VersionBump, Update Issue Templates, ChangeLog, Logo, Favicon.